norm check

+++ source/new_src/1-11/camera.c
@@ -6,42 +6,16 @@
 /*   By: sejpark <sejpark@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/05 22:41:59 by sejpark           #+#    #+#             */
-/*   Updated: 2021/03/03 16:53:25 by sejpark          ###   ########.fr       */
+/*   Updated: 2021/03/05 16:07:35 by sejpark          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "camera.h"
-#include "vec.h"
 #include <math.h>
 
-const float	g_pi = 3.1415926535897932385;
-/* *********************************** 수정 ********************************* */
-//t_camera	*ft_camera_set(t_camera *cam, t_point3 lookfrom, t_point3 lookat,
-//							t_vec3 vup, float vfov, float aspect_ratio)
-//{
-//	float		theta;
-//	float		h;
-//	float		viewport_height;
-//	float		viewport_width;
-//
-//	theta = ft_degrees_to_radians(vfov);
-//	h = tan(theta / 2);
-//	viewport_height = 2.0 * h;
-//	viewport_width = aspect_ratio * viewport_height;
-//	cam->w = ft_vec_unit_vec(ft_vec_sub(lookfrom, lookat));
-//	cam->u = ft_vec_unit_vec(ft_vec_cross(vup, cam->w));
-//	cam->v = ft_vec_cross(cam->w, cam->u);
-//	cam->origin = ft_vec_set_xyz(lookfrom.x, lookfrom.y, lookfrom.z);
-//	cam->horizontal = ft_vec_mul_f(viewport_width, cam->u);
-//	cam->vertical = ft_vec_mul_f(viewport_height, cam->v);
-//	cam->lower_left_corner = ft_vec_set_xyz(
-//		cam->origin.x - cam->horizontal.x / 2 - cam->vertical.x / 2 - cam->w.x,
-//		cam->origin.y - cam->horizontal.y / 2 - cam->vertical.y / 2 - cam->w.y,
-//		cam->origin.z - cam->horizontal.z / 2 - cam->vertical.z / 2 - cam->w.z);
-//	return (cam);
-//}
+const double	g_pi = 3.1415926535897932385;
 
-t_camera	ft_camera_set(t_point3 pos, t_vec3 dir, float vfov)
+t_camera	ft_camera_set(t_point3 pos, t_vec3 dir, double vfov)
 {
 	t_camera	cam;
 	t_cam_data	cam_data;
@@ -65,11 +39,13 @@ t_camera	ft_camera_set(t_point3 pos, t_vec3 dir, float vfov)
 		cam.origin.z - cam.horizontal.z / 2 - cam.vertical.z / 2 - cam.w.z);
 	return (cam);
 }
+
 /* ************************************************************************** */
 
 // 매개변수를 t_camera cam로 하면 render.c 파일의 ft_draw 함수에서 ft_camera_get_ray 함수를 호출할 때 첫번째 매개변수의 타입이 t_camera cam로 되어 있으므로 void*형의 content 변수를 t_camera *로 바꿔주고 역참조를 해야한다. 굳이 이렇게 보기힘들고 불편하기 할 필요가 없으므로 ft_camera_get_ray의 첫번째 매개변수를 t_camera *형으로 바꾸어서 ft_camera_get_ray를 호출할때 매개변수가 자동으로 void*형에서 t_camera*형으로 대입되게 한다.
-t_ray		ft_camera_get_ray(t_camera *cam, float u, float v)
-//t_ray		ft_camera_get_ray(t_camera cam, float u, float v)
+
+//t_ray		ft_camera_get_ray(t_camera cam, double u, double v)
+t_ray		ft_camera_get_ray(t_camera *cam, double u, double v)
 {
 	t_vec3	dir;
 
@@ -82,7 +58,7 @@ t_ray		ft_camera_get_ray(t_camera *cam, float u, float v)
 	return (ft_ray_set(cam->origin, ft_vec_unit_vec(dir)));
 }
 
-float		ft_degrees_to_radians(float degrees)
+double		ft_degrees_to_radians(double degrees)
 {
 	return (degrees * g_pi / 180.0);
 }
