+++ source/new_src/1-11/light.c
@@ -6,14 +6,14 @@
 /*   By: sejpark <sejpark@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/23 00:03:51 by sejpark           #+#    #+#             */
-/*   Updated: 2021/02/27 17:41:25 by sejpark          ###   ########.fr       */
+/*   Updated: 2021/03/05 16:11:34 by sejpark          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "light.h"
 #include <math.h>
 
-t_splight	ft_splight_set(t_point3 pos, float intensity, t_color color)
+t_splight	ft_splight_set(t_point3 pos, double intensity, t_color color)
 {
 	t_splight	li;
 
@@ -26,7 +26,7 @@ t_splight	ft_splight_set(t_point3 pos, float intensity, t_color color)
 t_spli_info	ft_splight_get_info(t_splight *spli, t_point3 *p)
 {
 	t_spli_info	spli_info;
-	float		r2;
+	double		r2;
 
 	spli_info.lightdir = ft_vec_sub(*p, spli->pos);
 	r2 = ft_vec_sqr_len(spli_info.lightdir);
@@ -37,7 +37,7 @@ t_spli_info	ft_splight_get_info(t_splight *spli, t_point3 *p)
 	return (spli_info);
 }
 
-t_dilight	ft_dilight_set(t_point3 pos, t_vec3 dir, float intensity,
+t_dilight	ft_dilight_set(t_point3 pos, t_vec3 dir, double intensity,
 					t_color color)
 {
 	t_dilight	dili;
