+++ source/new_src/1-11/triangle.h
@@ -6,7 +6,7 @@
 /*   By: sejpark <sejpark@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/21 20:32:51 by sejpark           #+#    #+#             */
-/*   Updated: 2021/03/01 20:07:24 by sejpark          ###   ########.fr       */
+/*   Updated: 2021/03/05 16:18:15 by sejpark          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,7 +43,7 @@ typedef struct	s_triangle
 t_triangle		ft_triangle_set(t_point3 p0, t_point3 p1, t_point3 p2,
 															t_color color);
 void			ft_triangle_set_c(t_triangle *tr, t_point3 p);
-float			ft_triangle_solve_t(t_triangle *tr, t_ray *r, t_t *t_range);
+double			ft_triangle_solve_t(t_triangle *tr, t_ray *r, t_t *t_range);
 int				ft_triangle_hit(t_triangle *tr, t_ray *r, t_t *t_range,
 															t_hit_rec *rec);
 int				ft_triangle_sha_hit(t_triangle *tr, t_ray *r, t_t *t_range);
